// src/generated/library/SearchResult.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Account } from "../live/Account";
import type { SearchGame } from "./SearchGame";

export type SearchResult = { account: Account, server: string, games: Array<SearchGame>, start: number, hasMore: boolean, skippedGames: number, source: string, };